GH-329 Enforce max attempt time
If the maximum time for an attempt is exceeded, do not even start to
execute the pre-select tasks but instead return immediately.
While most other things are implemented as pre-select tasks, in this
case it is easier to return early because we do not want to generate any
feedback for a question that would be selected and would skip the
remaining pre-select tasks anyway.