Fix: filter attempts in shortcode
In the catquizstatistics shortcode, the ability profile chart
incorrectly showed all attempts instead of no attempts if there were no
attempts for the given parameters. The reason was, that the function the
performed the DB query assumes that it should not filter by users if the
list of userids is empty. This is resolved now by skipping the DB query
if the userlist is empty.